+++ ApplicationCode/ReservoirDataModel/RigHexIntersectionTools.h
@@ -66,7 +66,7 @@ struct RigHexIntersectionTools
                                          std::list<std::pair<cvf::Vec3d, cvf::Vec3d > >& intersectionLineSegments);
 
     static bool planeHexIntersectionPolygons(std::array<cvf::Vec3d, 8> hexCorners,
-                                             cvf::Mat4f transformMatrixForPlane,
+                                             cvf::Mat4d transformMatrixForPlane,
                                              std::vector<std::vector<cvf::Vec3d> > & polygons);
 };
